+++ tests/tests.py
@@ -213,6 +213,39 @@ class MessageBirdBackendTests(BaseSmsBackendTests, SimpleTestCase):
         )
 
 
+class TwilioBackendTests(BaseSmsBackendTests, SimpleTestCase):
+    sms_backend = 'sms.backends.twilio.SmsBackend'
+
+    def setUp(self) -> None:
+        super().setUp()
+        self._settings_override = override_settings(
+            TWILIO_ACCOUNT_SID='fake_account_sid',
+            TWILIO_AUTH_TOKEN='fake_auth_token',
+        )
+        self._settings_override.enable()
+
+    def tearDown(self) -> None:
+        self._settings_override.disable()
+        super().tearDown()
+
+    def test_send_messages(self) -> None:
+        """Test send_messages with the Twilio backend."""
+        message = Message(
+            'Here is the message',
+            '+12065550100',
+            ['+441134960000']
+        )
+
+        connection = sms.get_connection()
+        connection.client.messages.create = MagicMock()  # type: ignore
+        connection.send_messages([message])  # type: ignore
+        connection.client.messages.create.assert_called_with(  # type: ignore
+            to='+441134960000',
+            from_='+12065550100',
+            body='Here is the message'
+        )
+
+
 class SignalTests(SimpleTestCase):
 
     def flush_mailbox(self) -> None: